pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prism"
version = "1.0.0"
description = "PRISM — Shadow AI Traffic Analyzer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "mitmproxy>=11.0.0",
    "playwright>=1.48.0",
    "scapy>=2.6.0",
    "dpkt>=1.9.8",
    "httptools>=0.6.1",
    "wsproto>=1.2.0",
    "h2>=4.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
    "cryptography>=43.0.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
postgres = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
prism = "prism.main:main"

[tool.hatch.build.targets.wheel]
packages = ["prism"]
